Added a workaround for ManualPosition not accepting two times the same value.
In my case i used Siri to close the Blinds. The next day i open
the Blinds using the Buttons in my House. The next time i'll use
Siri to close them again and this won't work.

The first time using Siri will result in a ManualPosition/100
Using my Buttons will result in something like FullUp in the API
The third action will result in a ManualPosition/100 again, which will
not work. ManualPosition has to get different value.

My Workaround is a low number 0.00xxxxxxxxxx which will be added to
ManualPosition. eg. ManualPosition/100.002334534773.